requirements.txt
numpy==2.4.0
opencv_python==4.12.0.88
Pillow==12.1.0
torch==2.9.1
torchvision==0.24.1
ultralytics==8.3.242
